Add new regexp based section block parser
Replace the old and basic section block parser with this slightly more
advanced one which works using regexp. It is in addition to parsing
simple section blocks also capable of handling quoted strings.

Previously having the field separator (,) inside a quoted string still
splitted the section block by the field separator and that is no longer
the case.

The quote characters (typically ") are still sent in the result, maybe
we should strip this before we send it back?